+++ src/main/java/vectorwing/farmersdelight/data/BlockTags.java
@@ -47,6 +47,11 @@ public class BlockTags extends BlockTagsProvider
 			ModBlocks.CABBAGE_CRATE.get(),
 			ModBlocks.TOMATO_CRATE.get(),
 			ModBlocks.ONION_CRATE.get(),
+			ModBlocks.APPLE_CRATE.get(),
+			ModBlocks.CHORUS_FRUIT_CRATE.get(),
+			ModBlocks.BROWN_MUSHROOM_CRATE.get(),
+			ModBlocks.RED_MUSHROOM_CRATE.get(),
+			ModBlocks.EGG_CARTON.get(),
 			ModBlocks.OAK_CABINET.get(),
 			ModBlocks.BIRCH_CABINET.get(),
 			ModBlocks.SPRUCE_CABINET.get(),
@@ -89,6 +94,9 @@ public class BlockTags extends BlockTagsProvider
 				Blocks.COBWEB,
 				Blocks.CAKE,
 				ModBlocks.RICE_BAG.get(),
+				ModBlocks.SWEET_BERRIES_BAG.get(),
+				ModBlocks.GLOW_BERRIES_BAG.get(),
+				ModBlocks.COCOA_BEANS_BAG.get(),
 				ModBlocks.APPLE_PIE.get(),
 				ModBlocks.SWEET_BERRY_CHEESECAKE.get(),
 				ModBlocks.CHOCOLATE_PIE.get(),
@@ -234,9 +242,16 @@ public class BlockTags extends BlockTagsProvider
 			CommonTags.Blocks.STORAGE_BLOCKS_CABBAGE,
 			CommonTags.Blocks.STORAGE_BLOCKS_TOMATO,
 			CommonTags.Blocks.STORAGE_BLOCKS_ONION,
-			CommonTags.Blocks.STORAGE_BLOCKS_RICE,
+			CommonTags.Blocks.STORAGE_BLOCKS_APPLE,
+			CommonTags.Blocks.STORAGE_BLOCKS_CHORUS_FRUIT,
+			CommonTags.Blocks.STORAGE_BLOCKS_BROWN_MUSHROOM,
+			CommonTags.Blocks.STORAGE_BLOCKS_RED_MUSHROOM,
+			CommonTags.Blocks.STORAGE_BLOCKS_STRAW,
 			CommonTags.Blocks.STORAGE_BLOCKS_RICE_PANICLE,
-			CommonTags.Blocks.STORAGE_BLOCKS_STRAW
+			CommonTags.Blocks.STORAGE_BLOCKS_RICE,
+			CommonTags.Blocks.STORAGE_BLOCKS_SWEET_BERRIES,
+			CommonTags.Blocks.STORAGE_BLOCKS_GLOW_BERRIES,
+			CommonTags.Blocks.STORAGE_BLOCKS_COCOA_BEANS
 		);
 	}
 
@@ -248,9 +263,16 @@ public class BlockTags extends BlockTagsProvider
 		tag(CommonTags.Blocks.STORAGE_BLOCKS_CABBAGE).add(ModBlocks.CABBAGE_CRATE.get());
 		tag(CommonTags.Blocks.STORAGE_BLOCKS_TOMATO).add(ModBlocks.TOMATO_CRATE.get());
 		tag(CommonTags.Blocks.STORAGE_BLOCKS_ONION).add(ModBlocks.ONION_CRATE.get());
-		tag(CommonTags.Blocks.STORAGE_BLOCKS_RICE).add(ModBlocks.RICE_BAG.get());
-		tag(CommonTags.Blocks.STORAGE_BLOCKS_RICE_PANICLE).add(ModBlocks.RICE_BALE.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_APPLE).add(ModBlocks.APPLE_CRATE.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_CHORUS_FRUIT).add(ModBlocks.CHORUS_FRUIT_CRATE.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_BROWN_MUSHROOM).add(ModBlocks.BROWN_MUSHROOM_CRATE.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_RED_MUSHROOM).add(ModBlocks.RED_MUSHROOM_CRATE.get());
 		tag(CommonTags.Blocks.STORAGE_BLOCKS_STRAW).add(ModBlocks.STRAW_BALE.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_RICE_PANICLE).add(ModBlocks.RICE_BALE.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_RICE).add(ModBlocks.RICE_BAG.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_SWEET_BERRIES).add(ModBlocks.SWEET_BERRIES_BAG.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_GLOW_BERRIES).add(ModBlocks.GLOW_BERRIES_BAG.get());
+		tag(CommonTags.Blocks.STORAGE_BLOCKS_COCOA_BEANS).add(ModBlocks.COCOA_BEANS_BAG.get());
 	}
 
 	protected void registerModTags() {